Eutester4J test refactorings and updates
