+++ eutester4j/com/eucalyptus/tests/awssdk/Eutester4j.java
@@ -39,7 +39,6 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.*;
 import java.util.concurrent.TimeUnit;
-import java.net.InetAddress;
 
 class Eutester4j {
