+++ eutester4j/com/eucalyptus/tests/awssdk/TestEC2VPCModifyInstanceSecurityGroups.groovy
@@ -10,7 +10,6 @@ import org.testng.annotations.Test;
 
 import static com.eucalyptus.tests.awssdk.Eutester4j.ACCESS_KEY
 import static com.eucalyptus.tests.awssdk.Eutester4j.EC2_ENDPOINT
-import static com.eucalyptus.tests.awssdk.Eutester4j.HOST_IP
 import static com.eucalyptus.tests.awssdk.Eutester4j.SECRET_KEY
 import static com.eucalyptus.tests.awssdk.Eutester4j.minimalInit
 
@@ -23,7 +22,6 @@ import static com.eucalyptus.tests.awssdk.Eutester4j.minimalInit
  */
 class TestEC2VPCModifyInstanceSecurityGroups {
 
-  private final String host;
   private final AWSCredentialsProvider credentials
 
   public static void main( String[] args ) throws Exception {
@@ -35,12 +33,6 @@ class TestEC2VPCModifyInstanceSecurityGroups {
 //
     this.credentials = new StaticCredentialsProvider( new BasicAWSCredentials( ACCESS_KEY, SECRET_KEY ) )
   }
-
-//  private String cloudUri( String servicePath ) {
-//    URI.create( "http://" + host + ":8773/" )
-//        .resolve( servicePath )
-//        .toString()
-//  }
 
   private AmazonEC2 getEC2Client( final AWSCredentialsProvider credentials ) {
     final AmazonEC2 ec2 = new AmazonEC2Client( credentials )
