+++ eutester4j/com/eucalyptus/tests/awssdk/TestEC2VPCStartStop.groovy
@@ -9,7 +9,7 @@ import com.amazonaws.services.ec2.model.*
 import org.testng.annotations.Test;
 
 import static com.eucalyptus.tests.awssdk.Eutester4j.ACCESS_KEY
-import static com.eucalyptus.tests.awssdk.Eutester4j.HOST_IP
+import static com.eucalyptus.tests.awssdk.Eutester4j.EC2_ENDPOINT
 import static com.eucalyptus.tests.awssdk.Eutester4j.SECRET_KEY
 import static com.eucalyptus.tests.awssdk.Eutester4j.minimalInit
 
@@ -22,7 +22,6 @@ import static com.eucalyptus.tests.awssdk.Eutester4j.minimalInit
  */
 class TestEC2VPCStartStop {
 
-  private final String host
   private final AWSCredentialsProvider credentials
   private final String cidrPrefix = '172.26.64'
 
@@ -32,14 +31,7 @@ class TestEC2VPCStartStop {
 
   public TestEC2VPCStartStop( ) {
     minimalInit()
-
     this.credentials = new StaticCredentialsProvider( new BasicAWSCredentials( ACCESS_KEY, SECRET_KEY ) )
-  }
-
-  private String cloudUri( String servicePath ) {
-    URI.create( "http://" + host + ":8773/" )
-        .resolve( servicePath )
-        .toString()
   }
 
   private AmazonEC2 getEC2Client( final AWSCredentialsProvider credentials ) {
